add predicate state, fix minor issues

+++ src/main/java/com/github/sadikovi/riff/tree/FilterApi.java
@@ -240,7 +240,7 @@ public class FilterApi {
       }
 
       @Override public DataType dataType() {
-        return DataTypes.IntegerType;
+        return DataTypes.LongType;
       }
 
       @Override public boolean evaluate(InternalRow row) {
